Clean up license bits
In this commit we add license headers to eucaconsole code that runs on
the server.  As this project contains code bundled from a great many
sources, per-file license statements on code original to this project
are especially important.

The files that contain third party license info now also contain
pseudo-file lists that enumerate the paths of the code they cover.
Ideally each project would reside in one or more dedicated directories.
That is usually the case in this code base, though there are exceptions.

Fixes GUI-281