Make setup.py install config files
setup.py's data_files directive doesn't appear to have an easy way
of renaming files as they are installed, so in addition to installing
our configuration files to /etc/eucaconsole, this commit also renames
conf/console.default.ini in the source tree to conf/console.ini.
Note that the "launcher.sh" script uses a configuration file at the
top of the source tree, not the conf directory.

Part of GUI-2869